refactor: 매일위키 api 관련 클래스 모듈 변경
