+++ maeilwiki-api/src/main/java/maeilwiki/auth/MemberIdentityCookieHelper.java
@@ -1,7 +1,7 @@
-package maeilwiki.member.api;
+package maeilwiki.auth;
 
-import static maeilwiki.member.api.MemberIdentityCookieType.ACCESS_TOKEN;
-import static maeilwiki.member.api.MemberIdentityCookieType.REFRESH_TOKEN;
+import static maeilwiki.auth.MemberIdentityCookieType.ACCESS_TOKEN;
+import static maeilwiki.auth.MemberIdentityCookieType.REFRESH_TOKEN;
 
 import jakarta.servlet.http.Cookie;
 import lombok.RequiredArgsConstructor;
