+++ maeilwiki-api/src/main/java/maeilwiki/auth/MemberIdentityExceptionHandler.java
@@ -1,4 +1,4 @@
-package maeilwiki.member.api;
+package maeilwiki.auth;
 
 import lombok.extern.slf4j.Slf4j;
 import maeilwiki.member.application.MemberIdentityException;
